run/helloworld/requirements.txt
streamlit
anthropic
google-cloud-secret-manager
